llm-api/requirements.txt
python-dotenv
ollama
openai
google-generativeai